[chore] reorganize the code
